moved help to its own plugin.
This is the first part of a refactoring that will move out all the
bot commands out of the base.backend-errbot-backend hierarchy.
